+++ src/hmf/_internals/_framework.py
@@ -207,6 +207,7 @@ class Framework(metaclass=_Validator):
     """
 
     _validate = True
+    _validate_every_param_set = False
 
     def validate(self):
         pass