+++ src/hmf/_internals/_cache.py
@@ -229,7 +229,7 @@ def parameter(kind):
                 and not isinstance(val, dict)
                 and val is not None
             ):
-                raise ValueError("%s must be a dictionary" % name)
+                raise ValueError(f"{name} must be a dictionary")
 
             # Locations of indexes
             recalc = hidden_loc(self, "recalc")
@@ -279,6 +279,18 @@ def parameter(kind):
                     # so need to re-index
                     for pr in getattr(self, recalc_papr)[name]:
                         delattr(self, pr)
+
+                if not doset and self._validate:
+                    if self._validate_every_param_set:
+                        self.validate()
+                    else:
+                        warnings.warn(
+                            f"You are setting {name} directly. This is unstable, as less "
+                            f"validation is performed. You can turn on extra validation "
+                            f"for directly set parameters by setting framework._validate_every_param_set=True."
+                            f"However, this can be brittle, since intermediate states may not be valid.",
+                            category=DeprecationWarning,
+                        )
 
         update_wrapper(_set_property, f)